GNOME - Fix crash on linux-arm64
After an account is created or opened, the app crashes with the
following console output:

	Registering new type NickvisionMoneyGNOMEViewsMainWindow with parent AdwApplicationWindow

	(NickvisionMoney.GNOME:95889): Gtk-WARNING **: 01:51:07.063: Theme parser error: <data>:1:61-62: Expected a valid selector
	Segmentation fault (core dumped)

GDB backtrace shows it crashes on the gtk_css_provider_load_from_data
function call:

	#43 0x0000ffbec8bdbbc0 in gtk_css_provider_load_from_data (css_provider=0xffbea80b8620,
	    data=0xffffdc9531d0 "#btnType { color: #fff; background-color: rgb(129,61,156); }", length=<optimized out>) at ../gtk/gtkcssprovider.c:1116
	#44 0x0000ffff74ad0af4 in ?? ()
	#45 0x0000ffffb7834360 in ?? () from /nix/store/h5l5d8xbx3prvzm51y6n5rvc5kajivcd-dotnet-runtime-7.0.0/shared/Microsoft.NETCore.App/7.0.0/libcoreclr.so
	Backtrace stopped: previous frame inner to this frame (corrupt stack?)

btnCss is 61 bytes long including the NUL terminator but it's reading
the 62-nd byte. I don't know why this happens and why it doesn't happen
on x86_64, but passing the length of btnCss fixes it.