Docs - Contributing.md, replace Peek with Kooha
I previously added Peek, because in the past I had great experience with this app, it was easy to use and its UI was awesome in pre-adwaita era. Unfortunately, not only it still has GTK3 UI, it turns out it doesn't support Wayland. I found Kooha and have been using it for some time, and tho I encountered some small issues, that's the app that I would recommend as easy-to-use video recorder app designed for GNOME.